added option to view all accounts, added a new menu for account management, and moved tests to match their class file
