+++ src/test/BankAccountTest.java
@@ -44,4 +44,48 @@ public class BankAccountTest {
             assertEquals(expected_transaction_history[i], transaction_history.get(i), 0.001);
         }
     }
+
+    @Test
+    public void testWithdraw() {
+        BankAccount testAccount = new BankAccount("checking", pin);
+        testAccount.deposit(100);
+        testAccount.withdraw(30);
+        assertEquals(70, testAccount.getBalance(), 0.01);
+    }
+
+    @Test
+    public void testInvalidWithdraw() {
+        BankAccount testAccount = new BankAccount("checking", pin);
+        testAccount.deposit(100);
+        try {
+            testAccount.withdraw(-50);
+            fail();
+        } catch (IllegalArgumentException e) {
+            // do nothing, test passes
+        }
+    }
+
+    @Test
+    public void testOverdraw() {
+        BankAccount testAccount = new BankAccount("checking", pin);
+        testAccount.deposit(60);
+        try {
+            testAccount.withdraw(70);
+            fail();
+        } catch (IllegalArgumentException e) {
+            // do nothing, test passes
+        }
+    }
+
+    @Test
+    public void testCheckBalance() {
+        BankAccount testAccount = new BankAccount("checking", "000000");
+        assertEquals(0, testAccount.getBalance(), 0.01);
+
+        testAccount.deposit(125);
+        assertEquals(125, testAccount.getBalance(), 0.01);
+
+        testAccount.withdraw(25);
+        assertEquals(100, testAccount.getBalance(), 0.01);
+    }
 }
